Use a better custom dict class
Directly subclassing dict is fraught with edge cases. For example, it's
the reason that we need to override both `__getitem__()` and also `get()` in
order to actually call `__getitem__()` in the first place.

collections.UserDict provides a good example base class based on
MutableMapping, which doesn't use magic C optimizations and actually
behaves as expected when overriding a method. It also conveniently
provides an attribute containing a real regular dict with the data --
we can use this for underlying lookups without invoking
`dict.__getitem__`.

Fixes an incompatibility with python 3.7 -- the custom `get()` function
uses 3.8-specific syntax from PEP 570: positional-only function
arguments.