Use python 3.7-compatible command joining code
`shlex.join()` was added in python 3.8, but this project specifies
support for 3.7. We cannot use this function (yet) but the good news is,
it is a simple convenience wrapper around string-joining a list of
`shlex.quote()`d arguments. It's straightforward to do the obvious thing
and get it to work.