Fix support for python 3.7 (#34)
* Use a better custom dict class

Directly subclassing dict is fraught with edge cases. For example, it's
the reason that we need to override both `__getitem__()` and also `get()` in
order to actually call `__getitem__()` in the first place.

collections.UserDict provides a good example base class based on
MutableMapping, which doesn't use magic C optimizations and actually
behaves as expected when overriding a method. It also conveniently
provides an attribute containing a real regular dict with the data --
we can use this for underlying lookups without invoking
`dict.__getitem__`.

Fixes an incompatibility with python 3.7 -- the custom `get()` function
uses 3.8-specific syntax from PEP 570: positional-only function
arguments.

* Use python 3.7-compatible command joining code

`shlex.join()` was added in python 3.8, but this project specifies
support for 3.7. We cannot use this function (yet) but the good news is,
it is a simple convenience wrapper around string-joining a list of
`shlex.quote()`d arguments. It's straightforward to do the obvious thing
and get it to work.

* CI: actually test on the minimum supported python version

Everything failed on python 3.7, despite that being the minimum
supported version listed. Start running CI tests to ensure this does not
regress.