ENH: support shared libraries inside packages
This drops the use of staged installs with `--destdir` in favor
of using `--prefix` for the install location. The problem is that
RPATHs in extension modules end up pointing to what Meson thinks is
the final install location (i.e., inside the prefix) rather than
to the staging directory.

`--destdir` seems meant for packaging, as an actual staging area,
while for `spin` the final install directory is `build-install`
(by default), there is no intent to later put this package into
`/usr` or `C:\`. Hence using `--prefix` seems like the correct
thing to do.

The one test change here is to a test that was incorrect.
`meson setup --prefix` expects an absolute path, and `/foobar`
isn't a path that exists or can be created.

Addresses the issue discussed in PR 238 - the `spin build` behavior
before this change cannot be made to work for SciPy, because the
internal shared library in `scipy.special` keeps breaking.

It should also address the problem discussed in issue spin#176.